requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
psycopg2-binary==2.9.7
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
networkx==3.1
python-dotenv==1.0.0
pydantic==2.4.2